src/requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
langchain-community
chromadb
sentence-transformers
tiktoken
